Refactor month list formatting function
